Feature/dropdown (#222)
* refactor(dropdown): break directive into separate files

* chore(angular): remove unused files

* docs(dropdown): update story

* chore(deps-dev): update

closes #221

* chore(segmented-control): rename angular story
