--- libs/angular/.storybook/preview-head.html
@@ -1,15 +1,0 @@
-<link
-  rel="stylesheet"
-  href="https://unpkg.com/@sebgroup/fonts@1.1.0/css/seb-fonts.min.css"
-/>
-<!-- Temporary fix since global styles aren't added correctly for angular -->
-<link
-  rel="stylesheet"
-  href="https://unpkg.com/@sebgroup/chlorophyll@latest/css/green-chlorophyll.min.css"
-/>
-<style>
-  body {
-    font-family: 'SEBSansSerif';
-    font-size: 16px;
-  }
-</style>
