--- libs/angular/.storybook/scss-loader.scss
@@ -1,6 +1,0 @@
-// load global styles used by storybook
-@use '../../libs/chlorophyll/src/lib/index.scss';
-body {
-  background: red;
-}
-
