+++ libs/angular/src/index.ts
@@ -1,7 +1,9 @@
-
 export * from './lib/green-angular.module'
 export * from './lib/dropdown/dropdown.module'
 export * from './lib/dropdown/dropdown.component'
 export * from './lib/dropdown/popover.directive'
+export * from './lib/dropdown/popover-element.directive'
+export * from './lib/dropdown/popover-option.directive'
+export * from './lib/dropdown/popover-trigger.directive'
 export * from './lib/segmented-control/segmented-control.module'
 export * from './lib/segmented-control/segmented-control.component'
