+++ libs/angular/src/lib/segmented-control/segmented-control.stories.ts
