+++ package.json
@@ -39,14 +39,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "12.2.5",
-    "@angular/common": "12.2.5",
-    "@angular/compiler": "12.2.5",
-    "@angular/core": "12.2.5",
-    "@angular/forms": "12.2.5",
-    "@angular/platform-browser": "12.2.5",
-    "@angular/platform-browser-dynamic": "12.2.5",
-    "@angular/router": "12.2.5",
+    "@angular/animations": "12.2.10",
+    "@angular/common": "12.2.10",
+    "@angular/compiler": "12.2.10",
+    "@angular/core": "12.2.10",
+    "@angular/forms": "12.2.10",
+    "@angular/platform-browser": "12.2.10",
+    "@angular/platform-browser-dynamic": "12.2.10",
+    "@angular/router": "12.2.10",
     "@popperjs/core": "^2.10.2",
     "@sebgroup/fonts": "^1.1.0",
     "browserslist": "^4.16.8",
@@ -59,12 +59,12 @@
     "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "12.2.5",
+    "@angular-devkit/build-angular": "12.1.4",
     "@angular-eslint/eslint-plugin": "^12.3.1",
     "@angular-eslint/eslint-plugin-template": "^12.3.1",
     "@angular-eslint/template-parser": "^12.3.1",
-    "@angular/compiler-cli": "12.2.5",
-    "@angular/language-service": "12.2.5",
+    "@angular/compiler-cli": "12.2.10",
+    "@angular/language-service": "12.2.10",
     "@babel/core": "^7.15.0",
     "@babel/plugin-proposal-private-property-in-object": "^7.14.5",
     "@babel/preset-typescript": "^7.15.0",
@@ -82,16 +82,16 @@
     "@nrwl/workspace": "12.9.0",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.0-rc.2",
     "@sebgroup/fonts": "^1.1.0",
-    "@storybook/addon-docs": "^6.4.0-beta.0",
-    "@storybook/addon-essentials": "^6.4.0-beta.0",
-    "@storybook/addon-links": "^6.4.0-beta.0",
+    "@storybook/addon-docs": "^6.4.0-beta.11",
+    "@storybook/addon-essentials": "^6.4.0-beta.11",
+    "@storybook/addon-links": "^6.4.0-beta.11",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/angular": "^6.4.0-beta.0",
-    "@storybook/builder-webpack5": "^6.4.0-beta.0",
-    "@storybook/html": "^6.4.0-beta.0",
-    "@storybook/manager-webpack5": "^6.4.0-beta.0",
-    "@storybook/react": "^6.4.0-beta.0",
-    "@storybook/theming": "^6.4.0-beta.0",
+    "@storybook/angular": "^6.4.0-beta.11",
+    "@storybook/builder-webpack5": "^6.4.0-beta.11",
+    "@storybook/html": "^6.4.0-beta.11",
+    "@storybook/manager-webpack5": "^6.4.0-beta.11",
+    "@storybook/react": "^6.4.0-beta.11",
+    "@storybook/theming": "^6.4.0-beta.11",
     "@svgr/webpack": "^5.5.0",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
